% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-solutions.R
\docType{data}
\name{aoc20_solutions}
\alias{aoc20_solutions}
\title{Advent of Code 2020 solutions}
\format{
An object of class \code{list} of length 50.
}
\usage{
aoc20_solutions
}
\description{
Advent of Code 2020 solutions
}
\keyword{datasets}
