% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{find_2020_pair}
\alias{find_2020_trio}
\alias{find_2020_trio_alternative}
\title{Day 01: Report Repair}
\usage{
find_2020_pair(x)

find_2020_trio(x)

find_2020_trio_alternative(x)
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{find_2020_pair(x)} returns the pair of numbers that
sum to 2020. For Part Two, \code{find_2020_trio(x)} returns the three numbers
that sum to 2020.
}
\description{
\href{https://adventofcode.com/2020/day/1}{Report Repair}
}
\details{
\strong{Part One}

After saving Christmas \href{/events}{five years in a row}, you\'ve decided
to take a vacation at a nice resort on a tropical island.
[Surely]{title="WHAT COULD GO WRONG"}, Christmas will go on without you.

The tropical island has its own currency and is entirely cash-only. The
gold coins used there have a little picture of a starfish; the locals
just call them \emph{stars}. None of the currency exchanges seem to have
heard of them, but somehow, you\'ll need to find fifty of these coins by
the time you arrive so you can pay the deposit on your room.

To save your vacation, you need to get all \emph{fifty stars} by December
25th.

Collect stars by solving puzzles. Two puzzles will be made available on
each day in the Advent calendar; the second puzzle is unlocked when you
complete the first. Each puzzle grants \emph{one star}. Good luck!

Before you leave, the Elves in accounting just need you to fix your
\emph{expense report} (your puzzle input); apparently, something isn\'t quite
adding up.

Specifically, they need you to \emph{find the two entries that sum to \code{2020}}
and then multiply those two numbers together.

For example, suppose your expense report contained the following:\preformatted{1721
979
366
299
675
1456
}

In this list, the two entries that sum to \code{2020} are \code{1721} and \code{299}.
Multiplying them together produces \code{1721 * 299 = 514579}, so the correct
answer is \code{514579}.

Of course, your expense report is much larger. \emph{Find the two entries
that sum to \code{2020}; what do you get if you multiply them together?}

\strong{Part Two}

The Elves in accounting are thankful for your help; one of them even
offers you a starfish coin they had left over from a past vacation. They
offer you a second one if you can find \emph{three} numbers in your expense
report that meet the same criteria.

Using the above example again, the three entries that sum to \code{2020} are
\code{979}, \code{366}, and \code{675}. Multiplying them together produces the answer,
\code{241861950}.

In your expense report, \emph{what is the product of the three entries that
sum to \code{2020}?}

\strong{Extra notes}

This problem is \href{https://en.wikipedia.org/wiki/3SUM}{3SUM} in disguise.
}
\examples{
x <- c(1721, 979, 366, 299, 675, 1456)
find_2020_pair(x)
x \%>\% find_2020_pair() \%>\% prod()
find_2020_trio(x)
}
