% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{count_trees_visited}
\title{Day 03: Toboggan Trajectory}
\usage{
count_trees_visited(x, move_x = 3, move_y = 1)
}
\arguments{
\item{x}{some data}

\item{move_x}{x position moved per step. Defaults to 3.}

\item{move_y}{y position moved per step. Defaults to 1.}
}
\value{
For Part One, \code{count_trees_visited(x)} returns the number of trees
visited by the sled. For Part Two, \code{f03b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2020/day/3}{Toboggan Trajectory}
}
\details{
\strong{Part One}

With the toboggan login problems resolved, you set off toward the
airport. While travel by toboggan might be easy, it\'s certainly not
safe: there\'s [very minimal
steering]{title="It looks like the toboggan steering system even runs on Intcode! Good thing you don't have to modify it."}
and the area is covered in trees. You\'ll need to see which angles will
take you near the fewest trees.

Due to the local geology, trees in this area only grow on exact integer
coordinates in a grid. You make a map (your puzzle input) of the open
squares (\code{.}) and trees (\verb{#}) you can see. For example:\preformatted{..##.......
#...#...#..
.#....#..#.
..#.#...#.#
.#...##..#.
..#.##.....
.#.#.#....#
.#........#
#.##...#...
#...##....#
.#..#...#.#
}

These aren\'t the only trees, though; due to something you read about
once involving arboreal genetics and biome stability, the same pattern
repeats to the right many times:\preformatted{..##.........##.........##.........##.........##.........##.......  --->
#...#...#..#...#...#..#...#...#..#...#...#..#...#...#..#...#...#..
.#....#..#..#....#..#..#....#..#..#....#..#..#....#..#..#....#..#.
..#.#...#.#..#.#...#.#..#.#...#.#..#.#...#.#..#.#...#.#..#.#...#.#
.#...##..#..#...##..#..#...##..#..#...##..#..#...##..#..#...##..#.
..#.##.......#.##.......#.##.......#.##.......#.##.......#.##.....  --->
.#.#.#....#.#.#.#....#.#.#.#....#.#.#.#....#.#.#.#....#.#.#.#....#
.#........#.#........#.#........#.#........#.#........#.#........#
#.##...#...#.##...#...#.##...#...#.##...#...#.##...#...#.##...#...
#...##....##...##....##...##....##...##....##...##....##...##....#
.#..#...#.#.#..#...#.#.#..#...#.#.#..#...#.#.#..#...#.#.#..#...#.#  --->
}

You start on the open square (\code{.}) in the top-left corner and need to
reach the bottom (below the bottom-most row on your map).

The toboggan can only follow a few specific slopes (you opted for a
cheaper model that prefers rational numbers); start by \emph{counting all the
trees} you would encounter for the slope \emph{right 3, down 1}:

From your starting position at the top-left, check the position that is
right 3 and down 1. Then, check the position that is right 3 and down 1
from there, and so on until you go past the bottom of the map.

The locations you\'d check in the above example are marked here with \code{O}
where there was an open square and \code{X} where there was a tree:\preformatted{..##.........##.........##.........##.........##.........##.......  --->
#..O#...#..#...#...#..#...#...#..#...#...#..#...#...#..#...#...#..
.#....X..#..#....#..#..#....#..#..#....#..#..#....#..#..#....#..#.
..#.#...#O#..#.#...#.#..#.#...#.#..#.#...#.#..#.#...#.#..#.#...#.#
.#...##..#..X...##..#..#...##..#..#...##..#..#...##..#..#...##..#.
..#.##.......#.X#.......#.##.......#.##.......#.##.......#.##.....  --->
.#.#.#....#.#.#.#.O..#.#.#.#....#.#.#.#....#.#.#.#....#.#.#.#....#
.#........#.#........X.#........#.#........#.#........#.#........#
#.##...#...#.##...#...#.X#...#...#.##...#...#.##...#...#.##...#...
#...##....##...##....##...#X....##...##....##...##....##...##....#
.#..#...#.#.#..#...#.#.#..#...X.#.#..#...#.#.#..#...#.#.#..#...#.#  --->
}

In this example, traversing the map using this slope would cause you to
encounter \code{7} trees.

Starting at the top-left corner of your map and following a slope of
right 3 and down 1, \emph{how many trees would you encounter?}

\strong{Part Two}

Time to check the rest of the slopes - you need to minimize the
probability of a sudden arboreal stop, after all.

Determine the number of trees you would encounter if, for each of the
following slopes, you start at the top-left corner and traverse the map
all the way to the bottom:
\itemize{
\item Right 1, down 1.
\item Right 3, down 1. (This is the slope you already checked.)
\item Right 5, down 1.
\item Right 7, down 1.
\item Right 1, down 2.
}

In the above example, these slopes would find \code{2}, \code{7}, \code{3}, \code{4}, and
\code{2} tree(s) respectively; multiplied together, these produce the answer
\code{336}.

\emph{What do you get if you multiply together the number of trees
encountered on each of the listed slopes?}
}
\examples{
x <- read_text_lines(
  "
  ..##.......
  #...#...#..
  .#....#..#.
  ..#.#...#.#
  .#...##..#.
  ..#.##.....
  .#.#.#....#
  .#........#
  #.##...#...
  #...##....#
  .#..#...#.#
  "
)
count_trees_visited(x, move_x = 3, move_y = 1)
count_trees_visited(x, move_x = 1, move_y = 2)
}
