% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{check_for_valid_passports}
\alias{check_for_valider_passports}
\title{Day 04: Passport Processing}
\usage{
check_for_valid_passports(x)

check_for_valider_passports(x)
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{check_for_valid_passports(x)} returns a logical vector
indicating whether each passport is valid. For Part Two, \code{f04b(x)} returns
....
}
\description{
\href{https://adventofcode.com/2020/day/4}{Passport Processing}
}
\details{
\strong{Part One}

You arrive at the airport only to realize that you grabbed your North
Pole Credentials instead of your passport. While these documents are
extremely similar, North Pole Credentials aren\'t issued by a country
and therefore aren\'t actually valid documentation for travel in most of
the world.

It seems like you\'re not the only one having problems, though; a very
long line has formed for the automatic passport scanners, and the delay
could upset your travel itinerary.

Due to some questionable network security, you realize you might be able
to solve both of these problems at the same time.

The automatic passport scanners are slow because they\'re having trouble
\emph{detecting which passports have all required fields}. The expected
fields are as follows:
\itemize{
\item \code{byr} (Birth Year)
\item \code{iyr} (Issue Year)
\item \code{eyr} (Expiration Year)
\item \code{hgt} (Height)
\item \code{hcl} (Hair Color)
\item \code{ecl} (Eye Color)
\item \code{pid} (Passport ID)
\item \code{cid} (Country ID)
}

Passport data is validated in batch files (your puzzle input). Each
passport is represented as a sequence of \code{key:value} pairs separated by
spaces or newlines. Passports are separated by blank lines.

Here is an example batch file containing four passports:\preformatted{ecl:gry pid:860033327 eyr:2020 hcl:#fffffd
byr:1937 iyr:2017 cid:147 hgt:183cm

iyr:2013 ecl:amb cid:350 eyr:2023 pid:028048884
hcl:#cfa07d byr:1929

hcl:#ae17e1 iyr:2013
eyr:2024
ecl:brn pid:760753108 byr:1931
hgt:179cm

hcl:#cfa07d eyr:2025 pid:166559648
iyr:2011 ecl:brn hgt:59in
}

The first passport is \emph{valid} - all eight fields are present. The second
passport is \emph{invalid} - it is missing \code{hgt} (the Height field).

The third passport is interesting; the \emph{only missing field} is \code{cid}, so
it looks like data from North Pole Credentials, not a passport at all!
Surely, nobody would mind if you made the system temporarily ignore
missing \code{cid} fields. Treat this \"passport\" as \emph{valid}.

The fourth passport is missing two fields, \code{cid} and \code{byr}. Missing
\code{cid} is fine, but missing any other field is not, so this passport is
\emph{invalid}.

According to the above rules, your improved system would report \code{2}
valid passports.

Count the number of \emph{valid} passports - those that have all required
fields. Treat \code{cid} as optional. \emph{In your batch file, how many passports
are valid?}

\strong{Part Two}

The line is moving more quickly now, but you overhear airport security
talking about how passports with invalid data are getting through.
Better add some data validation, quick!

You can continue to ignore the \code{cid} field, but each other field has
[strict rules]{title="GLORY TO ARSTOTZKA"} about what values are valid
for automatic validation:
\itemize{
\item \code{byr} (Birth Year) - four digits; at least \code{1920} and at most
\code{2002}.
\item \code{iyr} (Issue Year) - four digits; at least \code{2010} and at most
\code{2020}.
\item \code{eyr} (Expiration Year) - four digits; at least \code{2020} and at most
\code{2030}.
\item \code{hgt} (Height) - a number followed by either \code{cm} or \verb{in}:
\itemize{
\item If \code{cm}, the number must be at least \code{150} and at most \code{193}.
\item If \verb{in}, the number must be at least \code{59} and at most \code{76}.
}
\item \code{hcl} (Hair Color) - a \verb{#} followed by exactly six characters
\code{0}-\code{9} or \code{a}-\code{f}.
\item \code{ecl} (Eye Color) - exactly one of: \code{amb} \code{blu} \code{brn} \code{gry} \code{grn}
\code{hzl} \code{oth}.
\item \code{pid} (Passport ID) - a nine-digit number, including leading zeroes.
\item \code{cid} (Country ID) - ignored, missing or not.
}

Your job is to count the passports where all required fields are both
\emph{present} and \emph{valid} according to the above rules. Here are some
example values:\preformatted{byr valid:   2002
byr invalid: 2003

hgt valid:   60in
hgt valid:   190cm
hgt invalid: 190in
hgt invalid: 190

hcl valid:   #123abc
hcl invalid: #123abz
hcl invalid: 123abc

ecl valid:   brn
ecl invalid: wat

pid valid:   000000001
pid invalid: 0123456789
}

Here are some invalid passports:\preformatted{eyr:1972 cid:100
hcl:#18171d ecl:amb hgt:170 pid:186cm iyr:2018 byr:1926

iyr:2019
hcl:#602927 eyr:1967 hgt:170cm
ecl:grn pid:012533040 byr:1946

hcl:dab227 iyr:2012
ecl:brn hgt:182cm pid:021572410 eyr:2020 byr:1992 cid:277

hgt:59cm ecl:zzz
eyr:2038 hcl:74454a iyr:2023
pid:3556412378 byr:2007
}

Here are some valid passports:\preformatted{pid:087499704 hgt:74in ecl:grn iyr:2012 eyr:2030 byr:1980
hcl:#623a2f

eyr:2029 ecl:blu cid:129 byr:1989
iyr:2014 pid:896056539 hcl:#a97842 hgt:165cm

hcl:#888785
hgt:164cm byr:2001 iyr:2015 cid:88
pid:545766238 ecl:hzl
eyr:2022

iyr:2010 hgt:158cm hcl:#b6652a ecl:blu byr:1944 eyr:2021 pid:093154719
}

Count the number of \emph{valid} passports - those that have all required
fields \emph{and valid values}. Continue to treat \code{cid} as optional. \emph{In your
batch file, how many passports are valid?}
}
\examples{
x <-
  "
  ecl:gry pid:860033327 eyr:2020 hcl:#fffffd
  byr:1937 iyr:2017 cid:147 hgt:183cm

  iyr:2013 ecl:amb cid:350 eyr:2023 pid:028048884
  hcl:#cfa07d byr:1929

  hcl:#ae17e1 iyr:2013
  eyr:2024
  ecl:brn pid:760753108 byr:1931
  hgt:179cm

  hcl:#cfa07d eyr:2025 pid:166559648
  iyr:2011 ecl:brn hgt:59in
  "

x <- read_text_lines(x, drop_empty = "head/tail")
check_for_valid_passports(x)

x_bad <-
  "
  eyr:1972 cid:100
  hcl:#18171d ecl:amb hgt:170 pid:186cm iyr:2018 byr:1926

    iyr:2019
  hcl:#602927 eyr:1967 hgt:170cm
    ecl:grn pid:012533040 byr:1946

  hcl:dab227 iyr:2012
  ecl:brn hgt:182cm pid:021572410 eyr:2020 byr:1992 cid:277

  hgt:59cm ecl:zzz
  eyr:2038 hcl:74454a iyr:2023
  pid:3556412378 byr:2007
  "

x_good <-
  "
  pid:087499704 hgt:74in ecl:grn iyr:2012 eyr:2030 byr:1980
  hcl:#623a2f

  eyr:2029 ecl:blu cid:129 byr:1989
  iyr:2014 pid:896056539 hcl:#a97842 hgt:165cm

  hcl:#888785
  hgt:164cm byr:2001 iyr:2015 cid:88
  pid:545766238 ecl:hzl
  eyr:2022

  iyr:2010 hgt:158cm hcl:#b6652a ecl:blu byr:1944 eyr:2021 pid:093154719
  "

x_good <- read_text_lines(x_good, drop_empty = "head/tail")
x_bad <- read_text_lines(x_bad, drop_empty = "head/tail")

check_for_valider_passports(x_good)
check_for_valider_passports(x_bad)

}
