% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{convert_seat_string_to_number}
\title{Day 05: Binary Boarding}
\usage{
convert_seat_string_to_number(x)
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{f05a(x)} returns .... For Part Two,
\code{f05b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2020/day/5}{Binary Boarding}
}
\details{
\strong{Part One}

boarding pass! You aren\'t sure which seat is yours, and all of the
flight attendants are busy with the flood of people that suddenly made
it through passport control.

You write a [quick program]{title="No problem!"} to use your phone\'s
camera to scan all of the nearby boarding passes (your puzzle input);
perhaps you can find your seat through process of elimination.

Instead of \href{https://www.youtube.com/watch?v=oAHbLRjF0vo}{zones or groups}, this airline uses
\emph{binary space partitioning} to seat people. A seat might be specified
like \code{FBFBBFFRLR}, where \code{F} means \"front\", \code{B} means \"back\", \code{L}
means \"left\", and \code{R} means \"right\".

The first 7 characters will either be \code{F} or \code{B}; these specify exactly
one of the \emph{128 rows} on the plane (numbered \code{0} through \code{127}). Each
letter tells you which half of a region the given seat is in. Start with
the whole list of rows; the first letter indicates whether the seat is
in the \emph{front} (\code{0} through \code{63}) or the \emph{back} (\code{64} through \code{127}).
The next letter indicates which half of that region the seat is in, and
so on until you\'re left with exactly one row.

For example, consider just the first seven characters of \code{FBFBBFFRLR}:
\itemize{
\item Start by considering the whole range, rows \code{0} through \code{127}.
\item \code{F} means to take the \emph{lower half}, keeping rows \code{0} through \code{63}.
\item \code{B} means to take the \emph{upper half}, keeping rows \code{32} through \code{63}.
\item \code{F} means to take the \emph{lower half}, keeping rows \code{32} through \code{47}.
\item \code{B} means to take the \emph{upper half}, keeping rows \code{40} through \code{47}.
\item \code{B} keeps rows \code{44} through \code{47}.
\item \code{F} keeps rows \code{44} through \code{45}.
\item The final \code{F} keeps the lower of the two, \emph{row \code{44}}.
}

The last three characters will be either \code{L} or \code{R}; these specify
exactly one of the \emph{8 columns} of seats on the plane (numbered \code{0}
through \code{7}). The same process as above proceeds again, this time with
only three steps. \code{L} means to keep the \emph{lower half}, while \code{R} means to
keep the \emph{upper half}.

For example, consider just the last 3 characters of \code{FBFBBFFRLR}:
\itemize{
\item Start by considering the whole range, columns \code{0} through \code{7}.
\item \code{R} means to take the \emph{upper half}, keeping columns \code{4} through \code{7}.
\item \code{L} means to take the \emph{lower half}, keeping columns \code{4} through \code{5}.
\item The final \code{R} keeps the upper of the two, \emph{column \code{5}}.
}

So, decoding \code{FBFBBFFRLR} reveals that it is the seat at \emph{row \code{44},
column \code{5}}.

Every seat also has a unique \emph{seat ID}: multiply the row by 8, then add
the column. In this example, the seat has ID \code{44 * 8 + 5 = 357}.

Here are some other boarding passes:
\itemize{
\item \code{BFFFBBFRRR}: row \code{70}, column \code{7}, seat ID \code{567}.
\item \code{FFFBBBFRRR}: row \code{14}, column \code{7}, seat ID \code{119}.
\item \code{BBFFBBFRLL}: row \code{102}, column \code{4}, seat ID \code{820}.
}

As a sanity check, look through your list of boarding passes. \emph{What is
the highest seat ID on a boarding pass?}

\strong{Part Two}

\emph{Ding!} The \"fasten seat belt\" signs have turned on. Time to find your
seat.

It\'s a completely full flight, so your seat should be the only missing
boarding pass in your list. However, there\'s a catch: some of the seats
at the very front and back of the plane don\'t exist on this aircraft,
so they\'ll be missing from your list as well.

Your seat wasn\'t at the very front or back, though; the seats with IDs
+1 and -1 from yours will be in your list.

\emph{What is the ID of your seat?}
}
\examples{
x <- c("FBFBBFFRLR", "BFFFBBFRRR", "FFFBBBFRRR", "BBFFBBFRLL")
convert_seat_string_to_number(x)
f05b()
}
