% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06}
\alias{day06}
\alias{find_unique_questions}
\alias{find_shared_questions}
\title{Day 06: Custom Customs}
\usage{
find_unique_questions(x)

find_shared_questions(x)
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{f06a(x)} returns .... For Part Two,
\code{f06b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2020/day/6}{Custom Customs}
}
\details{
\strong{Part One}

much larger plane, \href{https://en.wikipedia.org/wiki/Customs_declaration}{customs declaration forms} are
distributed to the passengers.

The form asks a series of 26 yes-or-no questions marked \code{a} through \code{z}.
All you need to do is identify the questions for which \emph{anyone in your
group} answers \"yes\". Since your group is just you, this doesn\'t take
very long.

However, the person sitting next to you seems to be experiencing a
language barrier and asks if you can help. For each of the people in
their group, you write down the questions for which they answer \"yes\",
one per line. For example:\preformatted{abcx
abcy
abcz
}

In this group, there are \emph{\code{6}} questions to which anyone answered
\"yes\": \code{a}, \code{b}, \code{c}, \code{x}, \code{y}, and \code{z}. (Duplicate answers to the
same question don\'t count extra; each question counts at most once.)

Another group asks for your help, then another, and eventually you\'ve
collected answers from every group on the plane (your puzzle input).
Each group\'s answers are separated by a blank line, and within each
group, each person\'s answers are on a single line. For example:\preformatted{abc

a
b
c

ab
ac

a
a
a
a

b
}

This list represents answers from five groups:
\itemize{
\item The first group contains one person who answered \"yes\" to \emph{\code{3}}
questions: \code{a}, \code{b}, and \code{c}.
\item The second group contains three people; combined, they answered
\"yes\" to \emph{\code{3}} questions: \code{a}, \code{b}, and \code{c}.
\item The third group contains two people; combined, they answered \"yes\"
to \emph{\code{3}} questions: \code{a}, \code{b}, and \code{c}.
\item The fourth group contains four people; combined, they answered
\"yes\" to only \emph{\code{1}} question, \code{a}.
\item The last group contains one person who answered \"yes\" to only
\emph{\code{1}} question, \code{b}.
}

In this example, the sum of these counts is \code{3 + 3 + 3 + 1 + 1} =
\emph{\code{11}}.

For each group, count the number of questions to which anyone answered
\"yes\". \emph{What is the sum of those counts?}

\strong{Part Two}

As you finish the last group\'s customs declaration, you notice that
[you misread one
word]{title="Don't worry, nobody ever misreads just one word in real life."}
in the instructions:

You don\'t need to identify the questions to which \emph{anyone} answered
\"yes\"; you need to identify the questions to which \emph{everyone} answered
\"yes\"!

Using the same example as above:\preformatted{abc

a
b
c

ab
ac

a
a
a
a

b
}

This list represents answers from five groups:
\itemize{
\item In the first group, everyone (all 1 person) answered \"yes\" to
\emph{\code{3}} questions: \code{a}, \code{b}, and \code{c}.
\item In the second group, there is \emph{no} question to which everyone
answered \"yes\".
\item In the third group, everyone answered yes to only \emph{\code{1}} question,
\code{a}. Since some people did not answer \"yes\" to \code{b} or \code{c}, they
don\'t count.
\item In the fourth group, everyone answered yes to only \emph{\code{1}} question,
\code{a}.
\item In the fifth group, everyone (all 1 person) answered \"yes\" to
\emph{\code{1}} question, \code{b}.
}

In this example, the sum of these counts is \code{3 + 0 + 1 + 1 + 1} = \emph{\code{6}}.

For each group, count the number of questions to which \emph{everyone}
answered \"yes\". \emph{What is the sum of those counts?}
}
\examples{
x <- read_text_lines(
  drop_empty = "head/tail",
  x = "
  abc

  a
  b
  c

  ab
  ac

  a
  a
  a
  a

  b
  "
)
find_unique_questions(x)
find_shared_questions(x)
}
