% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day10.R
\name{day10}
\alias{day10}
\alias{count_adapter_differences}
\alias{count_adapter_orders}
\alias{example_adapter_data}
\title{Day 10: Adapter Array}
\usage{
count_adapter_differences(x)

count_adapter_orders(x)

example_adapter_data(example_number = 1)
}
\arguments{
\item{x}{some data}

\item{example_number}{which example data to use (1 or 2). Defaults to 1.}
}
\value{
For Part One, \code{count_adapter_differences()} returns the counts of
differences of 1, 2, 3 between adapters. For Part Two,
\code{count_adapter_orders(x)} returns the total number of adapter orders.
}
\description{
\href{https://adventofcode.com/2020/day/10}{Adapter Array}
}
\details{
\strong{Part One}

Patched into the aircraft\'s data port, you discover weather forecasts
of a massive tropical storm. Before you can figure out whether it will
impact your vacation plans, however, your device suddenly turns off!

Its battery is dead.

You\'ll need to plug it in. There\'s only one problem: the charging
outlet near your seat produces the wrong number of \emph{jolts}. Always
prepared, you make a list of all of the joltage adapters in your bag.

Each of your joltage adapters is rated for a specific \emph{output joltage}
(your puzzle input). Any given adapter can take an input \code{1}, \code{2}, or
\code{3} jolts \emph{lower} than its rating and still produce its rated output
joltage.

In addition, your device has a built-in joltage adapter rated for \emph{\code{3}
jolts higher} than the highest-rated adapter in your bag. (If your
adapter list were \code{3}, \code{9}, and \code{6}, your device\'s built-in adapter
would be rated for \code{12} jolts.)

Treat the charging outlet near your seat as having an effective joltage
rating of \code{0}.

Since you have some time to kill, you might as well test all of your
adapters. Wouldn\'t want to get to your resort and realize you can\'t
even charge your device!

If you \emph{use every adapter in your bag} at once, what is the distribution
of joltage differences between the charging outlet, the adapters, and
your device?

For example, suppose that in your bag, you have adapters with the
following joltage ratings:\preformatted{16
10
15
5
1
11
7
19
6
12
4
}

With these adapters, your device\'s built-in joltage adapter would be
rated for \code{19 + 3 = 22} jolts, 3 higher than the highest-rated adapter.

Because adapters can only connect to a source 1-3 jolts lower than its
rating, in order to use every adapter, you\'d need to choose them like
this:
\itemize{
\item The charging outlet has an effective rating of \code{0} jolts, so the
only adapters that could connect to it directly would need to have a
joltage rating of \code{1}, \code{2}, or \code{3} jolts. Of these, only one you
have is an adapter rated \code{1} jolt (difference of \emph{\code{1}}).
\item From your \code{1}-jolt rated adapter, the only choice is your \code{4}-jolt
rated adapter (difference of \emph{\code{3}}).
\item From the \code{4}-jolt rated adapter, the adapters rated \code{5}, \code{6}, or \code{7}
are valid choices. However, in order to not skip any adapters, you
have to pick the adapter rated \code{5} jolts (difference of \emph{\code{1}}).
\item Similarly, the next choices would need to be the adapter rated \code{6}
and then the adapter rated \code{7} (with difference of \emph{\code{1}} and \emph{\code{1}}).
\item The only adapter that works with the \code{7}-jolt rated adapter is the
one rated \code{10} jolts (difference of \emph{\code{3}}).
\item From \code{10}, the choices are \code{11} or \code{12}; choose \code{11} (difference of
\emph{\code{1}}) and then \code{12} (difference of \emph{\code{1}}).
\item After \code{12}, only valid adapter has a rating of \code{15} (difference of
\emph{\code{3}}), then \code{16} (difference of \emph{\code{1}}), then \code{19} (difference of
\emph{\code{3}}).
\item Finally, your device\'s built-in adapter is always 3 higher than the
highest adapter, so its rating is \code{22} jolts (always a difference of
\emph{\code{3}}).
}

In this example, when using every adapter, there are \emph{\code{7}} differences
of 1 jolt and \emph{\code{5}} differences of 3 jolts.

Here is a larger example:\preformatted{28
33
18
42
31
14
46
20
48
47
24
23
49
45
19
38
39
11
1
32
25
35
8
17
7
9
4
2
34
10
3
}

In this larger example, in a chain that uses all of the adapters, there
are \emph{\code{22}} differences of 1 jolt and \emph{\code{10}} differences of 3 jolts.

Find a chain that uses all of your adapters to connect the charging
outlet to your device\'s built-in adapter and count the joltage
differences between the charging outlet, the adapters, and your device.
\emph{What is the number of 1-jolt differences multiplied by the number of
3-jolt differences?}

\strong{Part Two}

To completely determine whether you have enough adapters, you\'ll need
to figure out how many different ways they can be arranged. Every
arrangement needs to connect the charging outlet to your device. The
previous rules about when adapters can successfully connect still apply.

The first example above (the one that starts with \code{16}, \code{10}, \code{15})
supports the following arrangements:\preformatted{(0), 1, 4, 5, 6, 7, 10, 11, 12, 15, 16, 19, (22)
(0), 1, 4, 5, 6, 7, 10, 12, 15, 16, 19, (22)
(0), 1, 4, 5, 7, 10, 11, 12, 15, 16, 19, (22)
(0), 1, 4, 5, 7, 10, 12, 15, 16, 19, (22)
(0), 1, 4, 6, 7, 10, 11, 12, 15, 16, 19, (22)
(0), 1, 4, 6, 7, 10, 12, 15, 16, 19, (22)
(0), 1, 4, 7, 10, 11, 12, 15, 16, 19, (22)
(0), 1, 4, 7, 10, 12, 15, 16, 19, (22)
}

(The charging outlet and your device\'s built-in adapter are shown in
parentheses.) Given the adapters from the first example, the total
number of arrangements that connect the charging outlet to your device
is \emph{\code{8}}.

The second example above (the one that starts with \code{28}, \code{33}, \code{18}) has
many arrangements. Here are a few:\preformatted{(0), 1, 2, 3, 4, 7, 8, 9, 10, 11, 14, 17, 18, 19, 20, 23, 24, 25, 28, 31,
32, 33, 34, 35, 38, 39, 42, 45, 46, 47, 48, 49, (52)

(0), 1, 2, 3, 4, 7, 8, 9, 10, 11, 14, 17, 18, 19, 20, 23, 24, 25, 28, 31,
32, 33, 34, 35, 38, 39, 42, 45, 46, 47, 49, (52)

(0), 1, 2, 3, 4, 7, 8, 9, 10, 11, 14, 17, 18, 19, 20, 23, 24, 25, 28, 31,
32, 33, 34, 35, 38, 39, 42, 45, 46, 48, 49, (52)

(0), 1, 2, 3, 4, 7, 8, 9, 10, 11, 14, 17, 18, 19, 20, 23, 24, 25, 28, 31,
32, 33, 34, 35, 38, 39, 42, 45, 46, 49, (52)

(0), 1, 2, 3, 4, 7, 8, 9, 10, 11, 14, 17, 18, 19, 20, 23, 24, 25, 28, 31,
32, 33, 34, 35, 38, 39, 42, 45, 47, 48, 49, (52)

(0), 3, 4, 7, 10, 11, 14, 17, 20, 23, 25, 28, 31, 34, 35, 38, 39, 42, 45,
46, 48, 49, (52)

(0), 3, 4, 7, 10, 11, 14, 17, 20, 23, 25, 28, 31, 34, 35, 38, 39, 42, 45,
46, 49, (52)

(0), 3, 4, 7, 10, 11, 14, 17, 20, 23, 25, 28, 31, 34, 35, 38, 39, 42, 45,
47, 48, 49, (52)

(0), 3, 4, 7, 10, 11, 14, 17, 20, 23, 25, 28, 31, 34, 35, 38, 39, 42, 45,
47, 49, (52)

(0), 3, 4, 7, 10, 11, 14, 17, 20, 23, 25, 28, 31, 34, 35, 38, 39, 42, 45,
48, 49, (52)
}

In total, this set of adapters can connect the charging outlet to your
device in \emph{\code{19208}} distinct arrangements.

You glance back down at your bag and try to remember why you brought so
many adapters; there must be \emph{more than a trillion} valid ways to
arrange them! Surely, there must be [an efficient
way]{title="Definitely itertools."} to count the arrangements.

\emph{What is the total number of distinct ways you can arrange the adapters
to connect the charging outlet to your device?}
}
\examples{
x <- example_adapter_data()
count_adapter_differences(x)
count_adapter_orders(x)
y <- example_adapter_data(2)
count_adapter_differences(y)
count_adapter_orders(y)
}
