% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day13.R
\name{day13}
\alias{day13}
\alias{estimate_earliest_bus}
\alias{estimate_earliest_shared_bus_time}
\alias{example_bus_notes}
\title{Day 13: Shuttle Search}
\usage{
estimate_earliest_bus(x)

estimate_earliest_shared_bus_time(x)

example_bus_notes(example = 1)
}
\arguments{
\item{x}{2-element character vector of bus notes.}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{estimate_earliest_bus(x)} returns the timestamp of the
earliest bus. For Part Two, \code{estimate_earliest_shared_bus_time(x)} returns
the time satisfying the constraints in the bus notes.
}
\description{
\href{https://adventofcode.com/2020/day/13}{Shuttle Search}
}
\details{
\strong{Part One}

Your ferry can make it safely to a nearby port, but it won\'t get much
further. When you call to book another ship, you discover that no ships
embark from that port to your vacation island. You\'ll need to get from
the port to the nearest airport.

Fortunately, a shuttle bus service is available to bring you from the
sea port to the airport! Each bus has an ID number that also indicates
\emph{how often the bus leaves for the airport}.

Bus schedules are defined based on a \emph{timestamp} that measures the
\emph{number of minutes} since some fixed reference point in the past. At
timestamp \code{0}, every bus simultaneously departed from the sea port.
After that, each bus travels to the airport, then various other
locations, and finally returns to the sea port to repeat its journey
forever.

The time this loop takes a particular bus is also its ID number: the bus
with ID \code{5} departs from the sea port at timestamps \code{0}, \code{5}, \code{10},
\code{15}, and so on. The bus with ID \code{11} departs at \code{0}, \code{11}, \code{22}, \code{33},
and so on. If you are there when the bus departs, you can ride that bus
to the airport!

Your notes (your puzzle input) consist of two lines. The first line is
your estimate of the \emph{earliest timestamp you could depart on a bus}. The
second line lists the bus IDs that are in service according to the
shuttle company; entries that show \code{x} must be out of service, so you
decide to ignore them.

To save time once you arrive, your goal is to figure out \emph{the earliest
bus you can take to the airport}. (There will be exactly one such bus.)

For example, suppose you have the following notes:\preformatted{939
7,13,x,x,59,x,31,19
}

Here, the earliest timestamp you could depart is \code{939}, and the bus IDs
in service are \code{7}, \code{13}, \code{59}, \code{31}, and \code{19}. Near timestamp \code{939},
these bus IDs depart at the times marked \code{D}:\preformatted{time   bus 7   bus 13  bus 59  bus 31  bus 19
929      .       .       .       .       .
930      .       .       .       D       .
931      D       .       .       .       D
932      .       .       .       .       .
933      .       .       .       .       .
934      .       .       .       .       .
935      .       .       .       .       .
936      .       D       .       .       .
937      .       .       .       .       .
938      D       .       .       .       .
939      .       .       .       .       .
940      .       .       .       .       .
941      .       .       .       .       .
942      .       .       .       .       .
943      .       .       .       .       .
944      .       .       D       .       .
945      D       .       .       .       .
946      .       .       .       .       .
947      .       .       .       .       .
948      .       .       .       .       .
949      .       D       .       .       .
}

The earliest bus you could take is bus ID \code{59}. It doesn\'t depart until
timestamp \code{944}, so you would need to wait \code{944 - 939 = 5} minutes
before it departs. Multiplying the bus ID by the number of minutes
you\'d need to wait gives \emph{\code{295}}.

\emph{What is the ID of the earliest bus you can take to the airport
multiplied by the number of minutes you\'ll need to wait for that bus?}

\strong{Part Two}

The shuttle company is running a
[contest]{title="This is why you should never let me design a contest for a shuttle company."}:
one gold coin for anyone that can find the earliest timestamp such that
the first bus ID departs at that time and each subsequent listed bus ID
departs at that subsequent minute. (The first line in your input is no
longer relevant.)

For example, suppose you have the same list of bus IDs as above:\preformatted{7,13,x,x,59,x,31,19
}

An \code{x} in the schedule means there are no constraints on what bus IDs
must depart at that time.

This means you are looking for the earliest timestamp (called \code{t}) such
that:
\itemize{
\item Bus ID \code{7} departs at timestamp \code{t}.
\item Bus ID \code{13} departs one minute after timestamp \code{t}.
\item There are no requirements or restrictions on departures at two or
three minutes after timestamp \code{t}.
\item Bus ID \code{59} departs four minutes after timestamp \code{t}.
\item There are no requirements or restrictions on departures at five
minutes after timestamp \code{t}.
\item Bus ID \code{31} departs six minutes after timestamp \code{t}.
\item Bus ID \code{19} departs seven minutes after timestamp \code{t}.
}

The only bus departures that matter are the listed bus IDs at their
specific offsets from \code{t}. Those bus IDs can depart at other times, and
other bus IDs can depart at those times. For example, in the list above,
because bus ID \code{19} must depart seven minutes after the timestamp at
which bus ID \code{7} departs, bus ID \code{7} will always \emph{also} be departing
with bus ID \code{19} at seven minutes after timestamp \code{t}.

In this example, the earliest timestamp at which this occurs is
\emph{\code{1068781}}:\preformatted{time     bus 7   bus 13  bus 59  bus 31  bus 19
1068773    .       .       .       .       .
1068774    D       .       .       .       .
1068775    .       .       .       .       .
1068776    .       .       .       .       .
1068777    .       .       .       .       .
1068778    .       .       .       .       .
1068779    .       .       .       .       .
1068780    .       .       .       .       .
1068781    D       .       .       .       .
1068782    .       D       .       .       .
1068783    .       .       .       .       .
1068784    .       .       .       .       .
1068785    .       .       D       .       .
1068786    .       .       .       .       .
1068787    .       .       .       D       .
1068788    D       .       .       .       D
1068789    .       .       .       .       .
1068790    .       .       .       .       .
1068791    .       .       .       .       .
1068792    .       .       .       .       .
1068793    .       .       .       .       .
1068794    .       .       .       .       .
1068795    D       D       .       .       .
1068796    .       .       .       .       .
1068797    .       .       .       .       .
}

In the above example, bus ID \code{7} departs at timestamp \code{1068788} (seven
minutes after \code{t}). This is fine; the only requirement on that minute is
that bus ID \code{19} departs then, and it does.

Here are some other examples:
\itemize{
\item The earliest timestamp that matches the list \verb{17,x,13,19} is
\emph{\code{3417}}.
\item \verb{67,7,59,61} first occurs at timestamp \emph{\code{754018}}.
\item \verb{67,x,7,59,61} first occurs at timestamp \emph{\code{779210}}.
\item \verb{67,7,x,59,61} first occurs at timestamp \emph{\code{1261476}}.
\item \verb{1789,37,47,1889} first occurs at timestamp \emph{\code{1202161486}}.
}

However, with so many bus IDs in your list, surely the actual earliest
timestamp will be larger than \code{100000000000000}!

\emph{What is the earliest timestamp such that all of the listed bus IDs
depart at offsets matching their positions in the list?}
}
\examples{
estimate_earliest_bus(example_bus_notes())
estimate_earliest_shared_bus_time(example_bus_notes())
}
