% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day15.R
\name{day15}
\alias{day15}
\alias{play_memory_game}
\alias{example_memory_game}
\title{Day 15: Rambunctious Recitation}
\usage{
play_memory_game(x, max_turn)

example_memory_game(example = 1)
}
\arguments{
\item{x}{the initial setup of the memory game}
}
\value{
For Part One, \code{play_memory_game(x)} returns a list of vectors with
game data. For Part Two, \code{f15b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2020/day/15}{Rambunctious Recitation}
}
\details{
\strong{Part One}

You catch the airport shuttle and try to book a new flight to your
vacation island. Due to the storm, all direct flights have been
cancelled, but a route is available to get around the storm. You take
it.

While you wait for your flight, you decide to check in with the Elves
back at the North Pole. They're playing a \emph{memory game} and are [ever
so excited]{title="Of course they are."} to explain the rules!

In this game, the players take turns saying \emph{numbers}. They begin by
taking turns reading from a list of \emph{starting numbers} (your puzzle
input). Then, each turn consists of considering the \emph{most recently
spoken number}:
\itemize{
\item If that was the \emph{first} time the number has been spoken, the current
player says \emph{\code{0}}.
\item Otherwise, the number had been spoken before; the current player
announces \emph{how many turns apart} the number is from when it was
previously spoken.
}

So, after the starting numbers, each turn results in that player
speaking aloud either \emph{\code{0}} (if the last number is new) or an \emph{age} (if
the last number is a repeat).

For example, suppose the starting numbers are \verb{0,3,6}:
\itemize{
\item \emph{Turn 1}: The \code{1}st number spoken is a starting number, \emph{\code{0}}.
\item \emph{Turn 2}: The \code{2}nd number spoken is a starting number, \emph{\code{3}}.
\item \emph{Turn 3}: The \code{3}rd number spoken is a starting number, \emph{\code{6}}.
\item \emph{Turn 4}: Now, consider the last number spoken, \code{6}. Since that was
the first time the number had been spoken, the \code{4}th number spoken
is \emph{\code{0}}.
\item \emph{Turn 5}: Next, again consider the last number spoken, \code{0}. Since it
\emph{had} been spoken before, the next number to speak is the difference
between the turn number when it was last spoken (the previous turn,
\code{4}) and the turn number of the time it was most recently spoken
before then (turn \code{1}). Thus, the \code{5}th number spoken is \code{4 - 1},
\emph{\code{3}}.
\item \emph{Turn 6}: The last number spoken, \code{3} had also been spoken before,
most recently on turns \code{5} and \code{2}. So, the \code{6}th number spoken is
\code{5 - 2}, \emph{\code{3}}.
\item \emph{Turn 7}: Since \code{3} was just spoken twice in a row, and the last two
turns are \code{1} turn apart, the \code{7}th number spoken is \emph{\code{1}}.
\item \emph{Turn 8}: Since \code{1} is new, the \code{8}th number spoken is \emph{\code{0}}.
\item \emph{Turn 9}: \code{0} was last spoken on turns \code{8} and \code{4}, so the \code{9}th
number spoken is the difference between them, \emph{\code{4}}.
\item \emph{Turn 10}: \code{4} is new, so the \code{10}th number spoken is \emph{\code{0}}.
}

(The game ends when the Elves get sick of playing or dinner is ready,
whichever comes first.)

Their question for you is: what will be the \emph{\code{2020}th} number spoken? In
the example above, the \code{2020}th number spoken will be \code{436}.

Here are a few more examples:
\itemize{
\item Given the starting numbers \verb{1,3,2}, the \code{2020}th number spoken is
\code{1}.
\item Given the starting numbers \verb{2,1,3}, the \code{2020}th number spoken is
\code{10}.
\item Given the starting numbers \verb{1,2,3}, the \code{2020}th number spoken is
\code{27}.
\item Given the starting numbers \verb{2,3,1}, the \code{2020}th number spoken is
\code{78}.
\item Given the starting numbers \verb{3,2,1}, the \code{2020}th number spoken is
\code{438}.
\item Given the starting numbers \verb{3,1,2}, the \code{2020}th number spoken is
\code{1836}.
}

Given your starting numbers, \emph{what will be the \code{2020}th number spoken?}

\strong{Part Two}

Impressed, the Elves issue you a challenge: determine the \code{30000000}th
number spoken. For example, given the same starting numbers as above:
\itemize{
\item Given \verb{0,3,6}, the \code{30000000}th number spoken is \code{175594}.
\item Given \verb{1,3,2}, the \code{30000000}th number spoken is \code{2578}.
\item Given \verb{2,1,3}, the \code{30000000}th number spoken is \code{3544142}.
\item Given \verb{1,2,3}, the \code{30000000}th number spoken is \code{261214}.
\item Given \verb{2,3,1}, the \code{30000000}th number spoken is \code{6895259}.
\item Given \verb{3,2,1}, the \code{30000000}th number spoken is \code{18}.
\item Given \verb{3,1,2}, the \code{30000000}th number spoken is \code{362}.
}

Given your starting numbers, \emph{what will be the \code{30000000}th number
spoken?}
}
\examples{
game <- play_memory_game(example_memory_game(1))
game$last_num
}
