% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day24.R
\name{day24}
\alias{day24}
\alias{flip_hex_tiles}
\alias{f24b}
\alias{example_hex_tiles}
\title{Day 24: Lobby Layout}
\usage{
flip_hex_tiles(x)

f24b(x)

example_hex_tiles(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{flip_hex_tiles(x)} returns the number of once visited
and twice visited tiles. For Part Two, \code{f24b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2020/day/24}{Lobby Layout}
}
\details{
\strong{Part One}

Your raft makes it to the tropical island; it turns out that the small
crab was an excellent navigator. You make your way to the resort.

As you enter the lobby, you discover a small problem: the floor is being
renovated. You can\'t even reach the check-in desk until they\'ve
finished installing the \emph{new tile floor}.

The tiles are all \emph{hexagonal}; they need to be arranged in a \href{https://en.wikipedia.org/wiki/Hexagonal_tiling}{hex grid} with a very
specific color pattern. Not in the mood to wait, you offer to help
figure out the pattern.

The tiles are all \emph{white} on one side and \emph{black} on the other. They
start with the white side facing up. The lobby is large enough to fit
whatever pattern might need to appear there.

A member of the renovation crew gives you a \emph{list of the tiles that need
to be flipped over} (your puzzle input). Each line in the list
identifies a single tile that needs to be flipped by giving a series of
steps starting from a \emph{reference tile} in the very center of the room.
(Every line starts from the same reference tile.)

Because the tiles are hexagonal, every tile has \emph{six neighbors}: east,
southeast, southwest, west, northwest, and northeast. These directions
are given in your list, respectively, as \code{e}, \code{se}, \code{sw}, \code{w}, \code{nw}, and
\code{ne}. A tile is identified by a series of these directions with \emph{no
delimiters}; for example, \code{esenee} identifies the tile you land on if
you start at the reference tile and then move one tile east, one tile
southeast, one tile northeast, and one tile east.

Each time a tile is identified, it flips from white to black or from
black to white. Tiles might be flipped more than once. For example, a
line like \code{esew} flips a tile immediately adjacent to the reference
tile, and a line like \code{nwwswee} flips the reference tile itself.

Here is a larger example:\preformatted{sesenwnenenewseeswwswswwnenewsewsw
neeenesenwnwwswnenewnwwsewnenwseswesw
seswneswswsenwwnwse
nwnwneseeswswnenewneswwnewseswneseene
swweswneswnenwsewnwneneseenw
eesenwseswswnenwswnwnwsewwnwsene
sewnenenenesenwsewnenwwwse
wenwwweseeeweswwwnwwe
wsweesenenewnwwnwsenewsenwwsesesenwne
neeswseenwwswnwswswnw
nenwswwsewswnenenewsenwsenwnesesenew
enewnwewneswsewnwswenweswnenwsenwsw
sweneswneswneneenwnewenewwneswswnese
swwesenesewenwneswnwwneseswwne
enesenwswwswneneswsenwnewswseenwsese
wnwnesenesenenwwnenwsewesewsesesew
nenewswnwewswnenesenwnesewesw
eneswnwswnwsenenwnwnwwseeswneewsenese
neswnwewnwnwseenwseesewsenwsweewe
wseweeenwnesenwwwswnew
}

In the above example, 10 tiles are flipped once (to black), and 5 more
are flipped twice (to black, then back to white). After all of these
instructions have been followed, a total of \emph{\code{10}} tiles are \emph{black}.

Go through the renovation crew\'s list and determine which tiles they
need to flip. After all of the instructions have been followed, \emph{how
many tiles are left with the black side up?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
flip_hex_tiles(example_hex_tiles())
f24b()
}
