% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functional.R
\name{find_value2}
\alias{find_value2}
\alias{find_value}
\alias{find_position}
\title{Find first item that matches certain conditions}
\usage{
find_value2(.x, .y, .p, ..., .dir = c("forward", "backward"), .default = NULL)

find_value(.x, .p, ..., .dir = c("forward", "backward"), .default = NULL)

find_position(.x, .p, ..., .dir = c("forward", "backward"), .default = NULL)
}
\arguments{
\item{.x, .y}{Vectors to iterate over. For 2-argument version, \code{.x}, \code{.y} must
have the same length.}

\item{.p}{A predicate function for determining matches.}

\item{...}{Additional arguments to set for the predicate}

\item{.dir}{Direction to traverse the items. One of \code{"forward"} (default) or
\code{"backward"}.}

\item{.default}{Value used when no match is found.}
}
\value{
either the \code{.default} value or the values of the first match.
}
\description{
The argument names here are meant to follow \link[purrr:detect]{purrr::detect}.
}
\examples{
find_value2(1:10, -4:5, function(x, y) x + y > 10)
find_value2(1:10, -4:5, function(x, y) x + y > 11)
find_value2(1:10, -4:5, function(x, y) x + y > 100)
}
