% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-functional.R
\name{lapply2}
\alias{lapply2}
\title{Map over two lists}
\usage{
lapply2(x, y, f, ...)
}
\arguments{
\item{x, y}{lists to apply a 2-argument function over.}

\item{f}{a function to apply.}

\item{...}{other arguments to the function.}
}
\value{
the result of applying the function to each pair of items in \code{x} and
\code{y}.
}
\description{
This function wraps over \code{\link[base:funprog]{base::Map()}} to create a replacement for
\code{\link[purrr:map2]{purrr::map2()}}.
}
