% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day12.R
\name{day12}
\alias{day12}
\alias{follow_ship_instructions}
\alias{follow_waypoint_instructions}
\alias{example_ship_instructions}
\title{Day 12: Rain Risk}
\usage{
follow_ship_instructions(x)

follow_waypoint_instructions(x)

example_ship_instructions()
}
\arguments{
\item{x}{ship instructions}
}
\value{
For Part One, \code{follow_ship_instructions(x)} returns the east/west
position, the north/south position and the angle of the ship (divided by
90). For Part Two, \code{follow_waypoint_instructions(x)} returns the waypoint
position followed by the ship position.
}
\description{
\href{https://adventofcode.com/2020/day/12}{Rain Risk}
}
\details{
\strong{Part One}

[faster than anyone
expected]{title="At least it wasn't a Category Six!"}. The ferry needs
to take \emph{evasive actions}!

Unfortunately, the ship\'s navigation computer seems to be
malfunctioning; rather than giving a route directly to safety, it
produced extremely circuitous instructions. When the captain uses the
\href{https://en.wikipedia.org/wiki/Public_address_system}{PA system} to ask
if anyone can help, you quickly volunteer.

The navigation instructions (your puzzle input) consists of a sequence
of single-character \emph{actions} paired with integer input \emph{values}. After
staring at them for a few minutes, you work out what they probably mean:
\itemize{
\item Action \emph{\code{N}} means to move \emph{north} by the given value.
\item Action \emph{\code{S}} means to move \emph{south} by the given value.
\item Action \emph{\code{E}} means to move \emph{east} by the given value.
\item Action \emph{\code{W}} means to move \emph{west} by the given value.
\item Action \emph{\code{L}} means to turn \emph{left} the given number of degrees.
\item Action \emph{\code{R}} means to turn \emph{right} the given number of degrees.
\item Action \emph{\code{F}} means to move \emph{forward} by the given value in the
direction the ship is currently facing.
}

The ship starts by facing \emph{east}. Only the \code{L} and \code{R} actions change
the direction the ship is facing. (That is, if the ship is facing east
and the next instruction is \code{N10}, the ship would move north 10 units,
but would still move east if the following action were \code{F}.)

For example:\preformatted{F10
N3
F7
R90
F11
}

These instructions would be handled as follows:
\itemize{
\item \code{F10} would move the ship 10 units east (because the ship starts by
facing east) to \emph{east 10, north 0}.
\item \code{N3} would move the ship 3 units north to \emph{east 10, north 3}.
\item \code{F7} would move the ship another 7 units east (because the ship is
still facing east) to \emph{east 17, north 3}.
\item \code{R90} would cause the ship to turn right by 90 degrees and face
\emph{south}; it remains at \emph{east 17, north 3}.
\item \code{F11} would move the ship 11 units south to \emph{east 17, south 8}.
}

At the end of these instructions, the ship\'s \href{https://en.wikipedia.org/wiki/Manhattan_distance}{Manhattan distance} (sum of the
absolute values of its east/west position and its north/south position)
from its starting position is \code{17 + 8} = \emph{\code{25}}.

Figure out where the navigation instructions lead. \emph{What is the
Manhattan distance between that location and the ship\'s starting
position?}

\strong{Part Two}

Before you can give the destination to the captain, you realize that the
actual action meanings were printed on the back of the instructions the
whole time.

Almost all of the actions indicate how to move a \emph{waypoint} which is
relative to the ship's position:
\itemize{
\item Action \emph{\code{N}} means to move the waypoint \emph{north} by the given value.
\item Action \emph{\code{S}} means to move the waypoint \emph{south} by the given value.
\item Action \emph{\code{E}} means to move the waypoint \emph{east} by the given value.
\item Action \emph{\code{W}} means to move the waypoint \emph{west} by the given value.
\item Action \emph{\code{L}} means to rotate the waypoint around the ship \emph{left}
(\emph{counter-clockwise}) the given number of degrees.
\item Action \emph{\code{R}} means to rotate the waypoint around the ship \emph{right}
(\emph{clockwise}) the given number of degrees.
\item Action \emph{\code{F}} means to move \emph{forward} to the waypoint a number of
times equal to the given value.
}

The waypoint starts \emph{10 units east and 1 unit north} relative to the
ship. The waypoint is relative to the ship; that is, if the ship moves,
the waypoint moves with it.

For example, using the same instructions as above:
\itemize{
\item \code{F10} moves the ship to the waypoint 10 times (a total of \emph{100 units
east and 10 units north}), leaving the ship at \emph{east 100, north 10}.
The waypoint stays 10 units east and 1 unit north of the ship.
\item \code{N3} moves the waypoint 3 units north to \emph{10 units east and 4 units
north of the ship}. The ship remains at \emph{east 100, north 10}.
\item \code{F7} moves the ship to the waypoint 7 times (a total of \emph{70 units
east and 28 units north}), leaving the ship at \emph{east 170, north 38}.
The waypoint stays 10 units east and 4 units north of the ship.
\item \code{R90} rotates the waypoint around the ship clockwise 90 degrees,
moving it to \emph{4 units east and 10 units south of the ship}. The ship
remains at \emph{east 170, north 38}.
\item \code{F11} moves the ship to the waypoint 11 times (a total of \emph{44 units
east and 110 units south}), leaving the ship at \emph{east 214, south
72}. The waypoint stays 4 units east and 10 units south of the ship.
}

After these operations, the ship\'s Manhattan distance from its starting
position is \code{214 + 72} = \emph{\code{286}}.

Figure out where the navigation instructions actually lead. \emph{What is the
Manhattan distance between that location and the ship\'s starting
position?}
}
\examples{
follow_ship_instructions(example_ship_instructions())
follow_waypoint_instructions(example_ship_instructions())
}
