% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day24.R
\name{day24}
\alias{day24}
\alias{flip_hex_tiles}
\alias{example_hex_tiles}
\title{Day 24: Lobby Layout}
\usage{
flip_hex_tiles(x)

example_hex_tiles(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{flip_hex_tiles(x)} returns the number of once visited
and twice visited tiles. For Part Two, \code{f24b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2020/day/24}{Lobby Layout}
}
\details{
\strong{Part One}

Your raft makes it to the tropical island; it turns out that the small
crab was an excellent navigator. You make your way to the resort.

As you enter the lobby, you discover a small problem: the floor is being
renovated. You can\'t even reach the check-in desk until they\'ve
finished installing the \emph{new tile floor}.

The tiles are all \emph{hexagonal}; they need to be arranged in a \href{https://en.wikipedia.org/wiki/Hexagonal_tiling}{hex grid} with a very
specific color pattern. Not in the mood to wait, you offer to help
figure out the pattern.

The tiles are all \emph{white} on one side and \emph{black} on the other. They
start with the white side facing up. The lobby is large enough to fit
whatever pattern might need to appear there.

A member of the renovation crew gives you a \emph{list of the tiles that need
to be flipped over} (your puzzle input). Each line in the list
identifies a single tile that needs to be flipped by giving a series of
steps starting from a \emph{reference tile} in the very center of the room.
(Every line starts from the same reference tile.)

Because the tiles are hexagonal, every tile has \emph{six neighbors}: east,
southeast, southwest, west, northwest, and northeast. These directions
are given in your list, respectively, as \code{e}, \code{se}, \code{sw}, \code{w}, \code{nw}, and
\code{ne}. A tile is identified by a series of these directions with \emph{no
delimiters}; for example, \code{esenee} identifies the tile you land on if
you start at the reference tile and then move one tile east, one tile
southeast, one tile northeast, and one tile east.

Each time a tile is identified, it flips from white to black or from
black to white. Tiles might be flipped more than once. For example, a
line like \code{esew} flips a tile immediately adjacent to the reference
tile, and a line like \code{nwwswee} flips the reference tile itself.

Here is a larger example:\preformatted{sesenwnenenewseeswwswswwnenewsewsw
neeenesenwnwwswnenewnwwsewnenwseswesw
seswneswswsenwwnwse
nwnwneseeswswnenewneswwnewseswneseene
swweswneswnenwsewnwneneseenw
eesenwseswswnenwswnwnwsewwnwsene
sewnenenenesenwsewnenwwwse
wenwwweseeeweswwwnwwe
wsweesenenewnwwnwsenewsenwwsesesenwne
neeswseenwwswnwswswnw
nenwswwsewswnenenewsenwsenwnesesenew
enewnwewneswsewnwswenweswnenwsenwsw
sweneswneswneneenwnewenewwneswswnese
swwesenesewenwneswnwwneseswwne
enesenwswwswneneswsenwnewswseenwsese
wnwnesenesenenwwnenwsewesewsesesew
nenewswnwewswnenesenwnesewesw
eneswnwswnwsenenwnwnwwseeswneewsenese
neswnwewnwnwseenwseesewsenwsweewe
wseweeenwnesenwwwswnew
}

In the above example, 10 tiles are flipped once (to black), and 5 more
are flipped twice (to black, then back to white). After all of these
instructions have been followed, a total of \emph{\code{10}} tiles are \emph{black}.

Go through the renovation crew\'s list and determine which tiles they
need to flip. After all of the instructions have been followed, \emph{how
many tiles are left with the black side up?}

\strong{Part Two}

The tile floor in the lobby is meant to be a [living art
exhibit]{title="I need one of these!"}. Every day, the tiles are all
flipped according to the following rules:
\itemize{
\item Any \emph{black} tile with \emph{zero} or \emph{more than 2} black tiles
immediately adjacent to it is flipped to \emph{white}.
\item Any \emph{white} tile with \emph{exactly 2} black tiles immediately adjacent
to it is flipped to \emph{black}.
}

Here, \emph{tiles immediately adjacent} means the six tiles directly touching
the tile in question.

The rules are applied \emph{simultaneously} to every tile; put another way,
it is first determined which tiles need to be flipped, then they are all
flipped at the same time.

In the above example, the number of black tiles that are facing up after
the given number of days has passed is as follows:\preformatted{Day 1: 15
Day 2: 12
Day 3: 25
Day 4: 14
Day 5: 23
Day 6: 28
Day 7: 41
Day 8: 37
Day 9: 49
Day 10: 37

Day 20: 132
Day 30: 259
Day 40: 406
Day 50: 566
Day 60: 788
Day 70: 1106
Day 80: 1373
Day 90: 1844
Day 100: 2208
}

After executing this process a total of 100 times, there would be
\emph{\code{2208}} black tiles facing up.

\emph{How many tiles will be black after 100 days?}
}
\examples{
flip_hex_tiles(example_hex_tiles())
}
